@@ -1,20 +1,19 @@
 {
   "name": "my-app",
   "version": "5.0.0",
-  "homepage": "https://gitname.github.io/react-gh-pages",
+  "homepage": "https://kesava.github.io/telugu-quiz",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.0",
     "@types/node": "^16.11.21",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "framer-motion": "^12.5.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-router-dom": "^6.30.0",
     "react-scripts": "5.0.0",
-    "typescript": "^4.5.4",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
@@ -44,6 +43,9 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^5.0.0"
+    "@types/react": "^19.0.10",
+    "@types/react-dom": "^19.0.4",
+    "gh-pages": "^5.0.0",
+    "typescript": "^5.8.2"
   }
 }
